pyproject.toml
[project]
name = "jobsentinel"
version = "0.1.0"
description = "Cybersecurity job automation system with AI-powered job matching"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "flask",
    "playwright>=1.57.0",
    "docker",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88